Improved stream compression
The `NBTi/oStream`s only provide a boolean flag to tell if the stream is compressed or not. In reality, there are three possible compressions: none, gzip or zlib. The latter one is heavily used to store Chunk NBT data in Minecraft.

This commit introduces three constants for the compression. They happen to be matching the ID convention used by Minecraft. This way, no additional case checking is required to handle the stream compression when reading region files.

The old methods using a boolean are marked as deprecated from now on, since they are ambiguous about which compression algorithm is meant.